165529388-(view parties): add parties container
