--- src/components/Layout.js
@@ -1,8 +1,0 @@
-import { Component } from 'react';
-import { connect } from 'react-redux';
-
-class Layout extends Component {
-
-}
-
-export default Layout;
